TimeTable의 Header의 Category 부분의 코드에서 컴포넌트를 분할했습니다.
